pyproject.toml
[project]
name = "ai-terminal"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.62.0",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyperclip>=1.11.0",
]
